requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==38.0.3
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.8
django-admin-honeypot @ https://github.com/dmpayton/django-admin-honeypot/archive/a840496d183df89965eeb33c9a5dd9ea3919dfff.zip
django-allauth==0.51.0
django-appconf==1.0.5
django-axes==5.39.0
django-bootstrap5==21.3
django-colorfield==0.7.2
django-compressor==4.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-csp==3.7
django-filter==22.1
django-ipware==4.0.2
django-jwt-oidc==0.3.9
django-libsass==0.9
django-password-validators==1.7.0
django-recaptcha==3.0.0
django-tables2==2.4.1
filetype==1.1.0
flake8==4.0.1
frozenlist==1.3.3
idna==3.3
jwcrypto==1.4
libsass==0.21.0
mccabe==0.6.1
multidict==6.0.3
oauthlib==3.2.2
Pillow==9.3.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pyflakes==2.4.0
PyJWT==2.6.0
python-http-client==3.3.7
python3-openid==3.2.0
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.16.0
slackclient==2.9.4
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
uuid==1.30
wrapt==1.14.1
yarl==1.8.2
